Nozomi.Base.Domain: ViewModels: Request: Minor changes for proper support

+++ Nozomi.Base.Domain/ViewModels/Request/CreateRequestViewModel.cs
@@ -29,7 +29,7 @@ namespace Nozomi.Data.ViewModels.Request
         /// <summary>
         /// The delay after a failed poll attempt
         /// </summary>
-        public int FailureDelay { get; set; }
+        public long FailureDelay { get; set; }
 
         /// <summary>
         /// This will deduce what type of request this is for
